PARTIAL IMPLEMENTATION
Insert slow path cloned loop blocks at end of EH region

This doesn't quite work for nested loops: nested loops in our current formulation
have the same kind of region structure and requirements as EH regions, namely,
a loop must be a contiguous range of blocks with a single entry block. We could move
the slow path out-of-line (and mark the blocks as not belonging to a loop -- another
problem). However, we need to branch back to the fast path / slow path join point, which
might be in the middle of a nested loop. This would create a potential 2nd entry to the
enclosing loop, which would break the required loop structure.

Instead of moving blocks to the end of the EH region, we could move them to the end of the enclosing loop.
However, that might not buy us much.

Possibly better would be a later-in-codegen pass to do code layout / straightening based
on block/edge weights. Since the slow path is already lightly weighted, it should naturally
fall out that it gets moved out of line. This phase wouldn't need to respect our loop structure
requirements, but it would still need to respect EH regions (unless we were able to generalize
EH regions as well).