Change pre-header creation for entry predecessors
Move all non-loop entry predecessor edges to point to the pre-header.
Before, only edges where the predecessor block was not dominated by
the entry would be moved.

This leads to a few diffs where some non-loop back-edges dominated by
the loop entry now point at the pre-header instead of the loop entry,
thus causing those paths to execute additional instructions that were
hoisted to the pre-header. This only occurs for cases where the non-loop
edge was not recognized as an outer loop.